three-d 0.7.3

+++ src/render/map_overlay.rs
@@ -14,8 +14,8 @@ impl Renderer<'_> {
         use three_d::vec3;
         use three_d::Viewport;
 
-        let overlay_width_px: u32 = 300;
-        let overlay_height_px: u32 = 300;
+        let overlay_width_px: usize = 300;
+        let overlay_height_px: usize = 300;
 
         let overlay_texture = ColorTargetTexture2D::<u8>::new(
             &self.context,
